requirements.txt
requests
tqdm
yadisk
python-dotenv
colorlog colorama